ref(installer): mount layout moves to disk/mount.py as LayoutMounter
* the teardown side already lived in disk/cleanup.py, the mount side did not
* mount() returns the fstab entries genfstab cannot see (encrypted swap)
